Refactor: 시간 설정 Controller 가 아닌 Service 에서 하도록 코드 수정
