+++ team2/src/main/java/com/team2/board/controller/BoardController.java
@@ -32,15 +32,12 @@ public class BoardController {
 	
 	@PostMapping(value="/create") 
 	public BoardVO testing(@RequestBody BoardVO board) {
-		board.setWriteDate(currentTime());
-		board.setUpdateDate(currentTime());
 		boardService.createBoard(board,boardService.maxBoardId());
 		return board;
 	}
 
 	@PostMapping(value="/createReply") 
 	public ReplyVO test2(@RequestBody ReplyVO reply) {
-		reply.setWriteDate(currentTime());
 		boardService.createReply(reply, boardService.maxReplyId());
 		return reply;
 	}
@@ -69,7 +66,6 @@ public class BoardController {
 
 	@PutMapping("/update")
 	public BoardVO putMethodName(@RequestBody BoardVO board) {
-		board.setUpdateDate(currentTime());
 		boardService.updateBoard(board);
 		return board;
 	}
@@ -174,13 +170,4 @@ public class BoardController {
     	boardService.createBoardTeam(boardTeamId);
     	return boardTeamId;
     }
-    
-	private Date currentTime() {
-        // 현재 날짜와 시간 얻기
-        java.util.Date currentDate = new java.util.Date();
-        // java.util.Date를 java.sql.Date로 변환
-        Date sqlDate = new Date(currentDate.getTime());
-        return sqlDate;
-	}
-
 }
